requirements.txt
appdirs==1.4.3
asgiref==3.2.7
asn1crypto==1.3.0
boto3==1.12.28
botocore==1.15.28
certifi==2019.11.28
cfe==0.0.15
cffi==1.14.0
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
django-s3direct==1.1.5
django-storages==1.9.1
django-summernote==0.8.11.6
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
packaging==20.3
pip-review==1.1.0
pipdate==0.4.1
psycopg2==2.8.4
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
pytube3==9.6.4
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
transliterate==1.10.2
typing-extensions==3.7.4.1
urllib3==1.25.8
whitenoise==5.0.1
